+++ cmd/checkfleet/main.go
@@ -18,6 +18,7 @@ import (
 	"time"
 
 	"github.com/Allan-Nava/checkfleet/internal/checks/certs"
+	"github.com/Allan-Nava/checkfleet/internal/checks/haproxy"
 	"github.com/Allan-Nava/checkfleet/internal/checks/httpcheck"
 	"github.com/Allan-Nava/checkfleet/internal/checks/nats"
 	"github.com/Allan-Nava/checkfleet/internal/engine"
@@ -47,7 +48,7 @@ func main() {
 
 func usage() {
 	fmt.Fprintln(os.Stderr, `uso:
-  checkfleet check <all|certs|http|nats> --config checkfleet.yml [--output text|markdown|json] [--exit-on-bad]
+  checkfleet check <all|certs|http|nats|haproxy> --config checkfleet.yml [--output text|markdown|json] [--exit-on-bad]
   checkfleet version`)
 }
 
@@ -92,6 +93,9 @@ func runCheck(args []string) error {
 		return err
 	}
 	if err := add("nats", func() engine.Check { return nats.New(*cfg.Checks.NATS) }, cfg.Checks.NATS != nil); err != nil {
+		return err
+	}
+	if err := add("haproxy", func() engine.Check { return haproxy.New(*cfg.Checks.HAProxy) }, cfg.Checks.HAProxy != nil); err != nil {
 		return err
 	}
 	if len(selected) == 0 {